Create VPC for all environments
Create a VPC for development, staging, and production.

We use the terraform VPC module. We host our VPC in the us-east-1 region
across multiple AZ. We have both public and private subnets, and set up
a nat gateway, so private subnets can access the internet.

The next step is creating a hardened bastion host so that we can access
hosts in our private subnet via ssh.