Demo provisioning machines via Ansible
Currently, the plan is to use Ansible for the one-time provisioning of
golden images via Packer. We choose Ansible because it has richer
abstractions than a shell script. We demo Ansible w/ a Vagrant builder.
Eventually, we will use build AMIs.

We create a `share` directory, for roles, boxes, scripts, etc... shared
across hosts.

We demo this in the name of creating a bastion host for which will be my
portal for ssh'ing into hosts in the VPC. This commit also includes some
notes around the bastion host.

In addition, we download `.box` files and store directly to decrease start up time of
download vagrant boxes.