Running Ansible on Packer
Testing running Ansible roles on Packer. Using the dev-sec.io linux
hardening role as a test.

We decided to utilize shared base images on which we will run common
roles. We will then create more individualized base images on top of
those. We will also have shared ansible roles which may or may not be
used in the shared base images.

We will store ansible roles in `packer/share` and they will be checked
into git at least to start.

The next step is determining how to execute inspec tests.