Run inspec tests while building images
Run the inspec tests on the images we build via Packer. We use inspec
because it appears to be the industry standard. We choose to run
`inspec` locally, so ensure `inspec` is installed on the machine.

Again, we decided to download/save the inspec profiles in source
control, even when they are from a 3rd party provider.