Finish vagrant image bastion
Get the vagrant image for the bastion host working. I've updated how we
import previous vagrant boxes. We need to be careful about the names
that we give boxes.

I've decided I want to run the inspec tests for all behaviors (both on
the base image and on the newer parts of the image). However, I will
_not_ run ansible roles that have already been run on base roles.
Mostly, I'm concerned about needing to copy role configuration into
multiple places.