Small updates to base AMI tags
Small updates to how we tag our base AMI. I decided that each base AMI
should have the full date in the title. We want to retain, within
reason, all copies of the AMIs we publish, so giving them unique names
will make them easier to work with.

Also, we want the `packer.json` for downstream AMI images to directly
specify the name of the base AMI from which they have been built.