[WIP] Add amazon-ebs builder for bastion host
The amazon-ebs builder will allow us to construct a AMI.

This builder is currently not working because we are unable to ssh into
the machine. My strong guess is that this is because of the
ssh-hardening done on the base image.

My first plan is to try creating a EC2 key pair and then using that to
log in. W.r.t. the key pair, I want to create it via generic terraform
(similar to how I will generically manage the `mattjmcnaughton.com`
DNS).