Make simple Terratest of `base` pass
Remove any of the AWS AMI specific checks... I'm not sure we really need
them.

Also, add a Makefile command for running the tests. Currently, the
single tests takes > 10 minutes, which is the default go timeout for
tests. So we need to specify a custom timeout.

Long term, we may want to give the go test the name `test` in the
Makefile.