Add playbook for generating wildcard ssl certs via Let's Encrypt and storing in S3
There are a number of different ways I could get the needed SSL certs
for mattjmcnaughton.com and mattjmcnaughton.io. Ultimately, I decided
the easiest way was generating wildcard certs using Let's Encrypt. We
use the DNS challenge so that we do not need to provision a separate
host just to serve the response to the challenge. We store the certs in
s3, and hosts which need them can download them. Eventually, we may swap
vault for s3.

We add a `orchestration` terraform environment with the necessary
one-off code for updating our TXT records with the contents of the
challenge.

We add a playbook describing the necessary steps to take. We will need
to refresh the certs every couple of months.

Also, fix a small terraform linting error.