Use asg w/ ELB for bastion host
Update our bastion host deployment to use an ASG, instead of launching
a single machine. Using an ASG makes us resilient to failure AND allows
us to use spot instances with minimal impact, which should decrease how
much we pay for hosts.

Given we're using an ASG, use an ELB to give a consistent public
DNS entry. With the use of the ELB, we can actually run the bastion ec2
instance in the private subnet, which keeps us from having any hosts on
the public subnet. Only load balancers are on our public subnets.

Note... the one downside of this plan is that ELB's are not
inexpensive (i.e. $18 per month). We could "potentially" address this
issue by replacing the ELB with a system to update a DNS entry whenever
the new bastion instance boots.

We will most likely use this pattern for all of our publicly accessible
web services.