Add firewall role
We add a `firewall` role, which configures our hosts to accept/deny
requests on certain ports. We will additionally use the cloud firewall
of whatever cloud provider we use (i.e. AWS Security Groups), but it
never hurts to have the additional protection. I made in effort to write
the firewall role in an os agnostic manner. Maybe I will one day deploy
BSD hosts in prod... it would be nice to have a consistent interface.

Attempting to add the `firewall` role exposed an issue with installing
`unattended-upgrade` in our base image. Whenever the base image boots,
so that we can apply additional provisioning to create our application
images, unattended-upgrades immediately runs and hogs the apt lock for
minutes. Ultimately, we decided uninstalling unattened-upgrades in the
base image, and reinstalling/configuring it in the application image, is
the easiest way to address this issue.

Finally, we add `ansible-lint` back, as we are comfortable installing
the dependency.

The next diff will be adding an nginx role.