Fix mattjmcnaughton.firewall role
Add some fixes for the firewall role - I noticed after booting to ec2
that it wasn't actually working how I expected. `ufw` didn't have the
settings I expected. Ultimately, the issue is I needed to be specifying
`role` instead of `policy`, as `policy` is a substitute for `default`.

I also got rid of the concept of a base image for both virtualbox and
AMI. A while back I was undecided between having a base image from which
we built our artifacts, and just running the same base playbooks on
everything. Ultimately, the `unattended-upgrades` bug with the `base`
image turned out to be difficult to address... so I'm going to fallback
to not having base artifacts, and instead just running the same base
provisioning on all of our application level artifacts. We can revisit
this decision once we have a couple of different artifacts.

I launched one of these bastion hosts via Terraform and verified I can
connect to it and that `ufw` is running.