Add mattjmcnaughton.nginx ansible role
Add an ansible role for running nginx on a host. Its completely tested
via inspec, and all tests are passing.

The most significant design decision relates to how we handle sensitive
information like the `.htpasswd` file and the ssl certs. We DO NOT want
to bake the production version of these into our AMI image. However, we
want to be able to test our configuration assuming those files exist. As
a result, our nginx role creates insecure "placeholders". We will
overwrite the placeholders when we actually launch the AMI. Importantly,
we will need to be sure that we do not serve any traffic (or count the
host as "ready") until the proper production creds have been put in
place.

One short coming is that for now, we only support one virtualhost per
machine. This restriction only exists for the MVP - we will want to
address in the future.