Add packer config for building blog artifacts
Add the packer config for building the blog artifacts. Doing so is
fairly simple, as we've already written the ansible playbook.

We updated the python we use on our remote machines for ansible to be `python3` and
also installed `python3-openssl`, which we need for generating the
self-signed certs.

Also, add a SCRATCH.md file outlining ideas for how we can install the
proper ssl certs on boot.