Update ASG to recreate when launch config changes
Set a config flag in our auto-scaling module so that when the launch
configuration changes, we create a new AMI. In practice, enabling this
option means that when we run `terraform apply` after creating a new
AMI, we will create a new launch configuration, AND a new auto-scaling
group, which means new hosts. The `terraform-aws-autoscaling` module
uses `create_before_destroy`, so all these changes should occur with
zero downtime.