[WIP] Provision k8s cluster via kubeadm
Provision a k8s cluster via kubeadm.

This commit does everything needed to use `kubeadm` on a node except run
`kubeadm init|join`. We will run those instructions on boot when we boot
the hosts (via a script we provision as part of ansible).

Our next step is to create AMIs via packer and test launching a master
and a worker, and ensuring we can get the two to communicate. We could
also use virtualbox images instead of a AMI...

We need to ensure said packer role has the correct firewall settings.