[WIP] Add packer for k8s master/workers
Add the packer for creating the images for master/worker k8s nodes.

Most importantly, we create a `integration/Vagrantfile` which will
launch a kubernetes cluster with one master and one worker and then test
that another machine can communicate with the worker. Instead of s3, we
use the file system to share `admin.conf`.

It's working which is exciting!

The next step is copying the changes from the `Vagrantfile` back to the
`kubeadm-init|join` scripts and then moving onto the Terraform.