Add mattjmcnaughton.docker role
Create a separate role for installing docker. We will use it for other
applications we plan on deploying (particularly in the time before we
have a working k8s cluster).

For now, our k8s role assumes this role has been run, however, we have
the flexibility to swap out docker for other container runtimes should
we desire.