[WIP] Add ansible, packer, terraform for vidzou
Add the ansible, packer, and terraform for vidzou. The deploy process
will be fairly similar to the blog.

TODOs/notes are captured in the source. I think the largest open
question is how I will determine the environment specific configurations
settings... I think the `env_file` in Docker could be useful.

First step is getting the `kitchen tests` for the ansible role written
and passing. We should check the app launches and is accessible via
nginx, although it obviously won't be able to access s3.