Finish packer for vidzou
Get packer working for vidzou. Uses the already tested ansible role. The
main thing we needed to do was some tweaking around packer running
ansible via python3, and also needing to run inspec tests as root, so
that we could call the docker api. We also need to set `inputs`, used by
the os-hardening tests, which allow the net forwarding sysctl (which is
needed by docker).