[WIP] Launch vidzou via terraform
WIP for launching vidzou via terraform! Largely based of the `blog`.
Definitely the chance to resolve some code duplication down the line.

A couple notes... we need to use a TCP health check, because HTTPS
health checks fail bc of auth.

Also, not completely working bc we need to configure our application to
read its configuration via a yaml file, because environment variables
are not updated when the container restarts.