Refactor to extract base resources
When we tried to create `vidzou` and `blog` together, we ran into issues
because both were attempting to define certain "base" security groups
and IAM policies. When we tried to create both, we got naming conflicts.

We can address this issue by creating a `base` module which will define
these entities, and then "inject" them into the application modules.

Over time, we find even more that we can extract into the base module.