Add real htaccess for vidzou
Similar to w/ ssl certs, we want to replace our "test" htpasswd file
with the real one when we deploy to the cloud.

Follow a similar approach to the one we use for ssl certs.

For now, we don't have separate `htpasswd` files for dev/prod... should
we find the need, we can considerate. But I'm not super worried about it
right now.