Implement a number of measures to save AWS bill
This diff contains a number of measures to reduce AWS costs.

Most importantly, we do not create a private subnet w/ a NAT gateway,
as that's extremely expensive. As additional fallout, we don't need a
bastion host anymore, so we delete that module.

The largest impact is that all ec2 instances are now exposed to the
public internet. We're relying on the security privated by our ssh
private key (we don't allow password auth).

We're not necessarily done w/ cost-savings. I'd love to stop paying for
two elbs.