Migrate to use single ALB for all applications instead of per-application ELBs
Previously, we were giving each application its own dedicated ELB. Given
an ELB costs ~$20 a month, we were paying at minimum of $20x per month,
where `x` is the number of applications.

If we use an ALB, than we can route to specific collections of hosts
based on the hostname. This allows us to use a single ALB across all
applications. Its a flat cost of ~$20, instead of $20 per application.

We create the `alb` as part of our `base` module, and then inject it
into the `blog` and `vidzou` modules. The `blog` and `vidzou` modules
can add the rules to perform the proper routing and register their
instances w/ the appropriate target groups.

One note - to use an application load balancer, we must terminate SSL at
the ALB. We upload our `mattjmcnaughton.com` certs to Amazon's ACM. For
now, we will still generate using the old Let's Encrypt process - we
will just have an additional place we need to upload them.

Note - in the long term, we should have the `global` environment read
the SSL certs from S3 instead of our local machine.