Remove load balancer and Auto-scaling groups
I can save ~50% on the AWS bill if I stop using ASGs and a Load
Balancer, and instead just launch EC2 records w/ A records to their
public IP.

Since I only have a single website (the blog), this should be fine.

I also deleted the `vidzou` module as its no longer in use, and will no
longer work as a result of getting rid of the lb. If we reintroduce it,
we will need to refactor to work without an lb.

Finally, stop managing the ACM certs via the `global` environment, as
only the LB used the ACM. The ec2 instances themselves pull the certs
from s3 on boot.